[ALPS09686149] cam: is8s coverity
[Description]
fix coverity

[Test]
pdas

MTK-Commit-Id: 1d5f7487b8e521fa01a1c1f5c5bc16b6bdd39a8b

CR-Id: ALPS09686149
Feature: [Android Default] Camera Application Basic Functions

+++ camsys/isp8s/cam/mtk_cam-raw_pads.h
@@ -49,7 +49,7 @@ static inline bool raw_is_valid_pad(unsigned int id)
 static inline int raw_pad_to_node_idx(unsigned int id)
 {
 	WARN_ON(id < MTK_RAW_SINK_NUM);
-	return id - MTK_RAW_SINK_NUM;
+	return (int)id - MTK_RAW_SINK_NUM;
 }
 
 #define _raw_pad_name(PAD)	\
